+++ unkey/routes.py
@@ -81,7 +81,8 @@ class Route:
 
 # Keys
 CREATE_KEY: t.Final[Route] = Route(c.POST, "/keys.createKey")
-VERIFY_KEY: t.Final[Route] = Route(c.POST, "/keys/verify")
+VERIFY_KEY: t.Final[Route] = Route(c.POST, "/keys.verifyKey")
+VERIFY_KEY_DEPRECATED: t.Final[Route] = Route(c.POST, "/keys/verify")
 REVOKE_KEY: t.Final[Route] = Route(c.DELETE, "/keys/{}")
 UPDATE_KEY: t.Final[Route] = Route(c.PUT, "/keys/{}")